proxy/requirements.txt
fastapi
uvicorn[standard]
httpx
boto3
PyJWT[crypto]
cryptography
requests
python-multipart
